Componentai israso duomenis is masyvo
