+++ src/App.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import './style/style.css';
 
-import Header from './components/pages/Header';
-import Content from './components/pages/Content';
+import Header from './components/includes/Header';
+import Content from './components/includes/Content';
 
 class App extends Component {
