--- src/components/Imones.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-
-class Imones extends Component {
-
-
-  render() {
-    return (
-      <div >
-        <h1>Imones</h1>
-      </div>
-    );
-  }
-}
-
-export default Imones;
