--- src/components/constants/Array.js
@@ -1,24 +1,0 @@
-import React from 'react';
-
-const imones = () => {
-[
-  {
-    id: 1,
-    title: sijonai
-  },
-  {
-    id: 2,
-    title: balvonai
-  },
-  {
-    id: 3,
-    title: klijai
-  },
-  {
-    id: 4,
-    title: kelnes
-  }
-]
-}
-
-export default imones;
