+++ src/components/constants/ImonesArray.js
@@ -1,0 +1,25 @@
+
+
+const ImonesArray = {
+imones:
+[
+  {
+    id: 1,
+    title: "sijonai"
+  },
+  {
+    id: 2,
+    title: "balvonai"
+  },
+  {
+    id: 3,
+    title: "klijai"
+  },
+  {
+    id: 4,
+    title: "kelnes"
+  }
+]
+}
+
+export default ImonesArray;
