+++ src/components/includes/Content.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { Route } from 'react-router';
 
-import Imones from '../Imones';
-import Index from '../Index';
+import Imones from '../pages/Imones';
+import Index from '../pages/Index';
 
 class Content extends Component {
