+++ src/components/pages/Imones.js
@@ -1,0 +1,36 @@
+import React, { Component } from 'react';
+import ContentItem from './ContentItem';
+
+class Imones extends Component {
+
+state = {
+  imones:
+  [
+    {
+      id: 1,
+      title: "sijonai"
+    },
+    {
+      id: 2,
+      title: "balvonai"
+    },
+    {
+      id: 3,
+      title: "klijai"
+    },
+    {
+      id: 4,
+      title: "kelnes"
+    }
+  ]
+}
+
+
+  render() {
+    return  this.state.imones.map((record) => (
+      <ContentItem key={record.id} item={record}/>
+    ));
+  }
+}
+
+export default Imones;
