+++ src/style/style.css
@@ -14,3 +14,33 @@
    color: #333;
    text-decoration: none;
  }
+
+.header {
+  font-family: 'Alex Brush', cursive, 'Cookie';
+  height:30%;
+  font-size: 2em;
+  font-weight: normal;
+  color: #FFD48b;
+  text-align:center;
+}
+
+.navigation-container {
+  width: 100%;
+  position: relative;
+  overflow: hidden;
+  text-align: center;
+}
+
+.top-nav {
+  display: inline-flex;
+  font-size: 1em;
+  text-decoration: none;
+  text-transform: uppercase;
+  font-weight: 400;
+  list-style: none;
+}
+
+.top-nav li {
+  padding: 0.2em;
+  border:black solid;
+}
